another item-action: 'i' to #adjust inventory
Add context-sensitive inventory item-action 'i' to run the #adjust
command.  Groundwork for item-action 'I' to use #adjust to split a
stack is included but non-functional.

Also reorders some of the enums and switch cases to be alphabetical
by their menu selection letter, dip moves from 'd' (which is drop)
to 'a' and offer/sacrifice moves from last to 'O'.