more shop wall repair
Commit 91aad708ecc9b7debdbf5e94acc668c9702f3585 took the 'catchup'
argument away from repair_damage() and calculated it in the routine.
Commit 89382aa83d4067934c695f6d8b95f62913b1f0c2 put the argument
back but neglected to remove calculating and assigning it, making
the passed value be ignored.  Take that away, finishing 89382aa83d.

That affected shop repair messages but wasn't enough to prevent a
shop wall repair display glitch.  This seems to make things work
properly but is little iffy.