more shop wall repair
Commit b1f1b1fe9f6d5e55b6c24b90acf4112a03752335 took the 'catchup'
argument away from repair_damage() and calculated it in the routine.
Commit c9068bf016ac2e20a63da120811f9caea071052c put the argument
back but neglected to remove calculating and assigning it, making
the passed value be ignored.  Take that away, finishing c9068bf016.

That affected shop repair messages but wasn't enough to prevent a
shop wall repair display glitch.  This seems to make things work
properly but is little iffy.